優化專案結構

+++ public/js/backend/admin/upload.js
@@ -25,26 +25,58 @@ function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o =
 function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }
 var droptarget = document.getElementById('drop');
 var dropImageList = document.getElementById('dropImageList');
+var imageQuantity = document.querySelector('.count');
+var uploadButton = document.querySelector('#uploadButton');
 var fileArr = [];
 var fileBlodArr = [];
-function handleEvent(event) {
-  event.preventDefault();
-  if (event.type === 'drop') {
-    droptarget.classList.remove('active');
-    var _iterator = _createForOfIteratorHelper(event.dataTransfer.files),
-      _step;
-    try {
-      for (_iterator.s(); !(_step = _iterator.n()).done;) {
-        var file = _step.value;
+uploadButton.addEventListener('change', function (event) {
+  var files = event.target.files;
+
+  // 遍历文件列表，将文件添加到 fileArr 数组中
+  var _iterator = _createForOfIteratorHelper(files),
+    _step;
+  try {
+    for (_iterator.s(); !(_step = _iterator.n()).done;) {
+      var file = _step.value;
+      if (fileArr.length < 10) {
         fileArr.push(file);
         filesToBlod(file);
+        droptarget.classList.remove('disable');
+        document.querySelector('.note').innerText = "Add Image";
+      } else {
+        droptarget.classList.add('disable');
+        document.querySelector('.note').innerText = "Max";
+      }
+    }
+
+    // 重置 input 元素的值，清空选择的文件
+  } catch (err) {
+    _iterator.e(err);
+  } finally {
+    _iterator.f();
+  }
+  event.target.value = '';
+});
+function handleEvent(event) {
+  event.preventDefault();
+  if (event.type === 'drop' || event.type === 'click') {
+    droptarget.classList.remove('active');
+    var _iterator2 = _createForOfIteratorHelper(event.dataTransfer.files),
+      _step2;
+    try {
+      for (_iterator2.s(); !(_step2 = _iterator2.n()).done;) {
+        var file = _step2.value;
+        if (fileArr.length < 10) {
+          fileArr.push(file);
+          filesToBlod(file);
+        }
       }
     } catch (err) {
-      _iterator.e(err);
+      _iterator2.e(err);
     } finally {
-      _iterator.f();
+      _iterator2.f();
     }
-  } else if (event.type === 'dropleave') {
+  } else if (event.type === 'dragleave') {
     droptarget.classList.remove('active');
   } else {
     droptarget.classList.add('active');
@@ -56,54 +88,80 @@ droptarget.addEventListener('dragenter', handleEvent);
 droptarget.addEventListener('dragover', handleEvent);
 droptarget.addEventListener('drop', handleEvent);
 droptarget.addEventListener('dragleave', handleEvent);
+droptarget.addEventListener('click', function () {
+  if (fileArr.length < 10) {
+    var _uploadButton = document.getElementById('uploadButton');
+    _uploadButton.click();
+    droptarget.classList.remove('disable');
+    document.querySelector('.note').innerText = "Add Image";
+  } else {
+    droptarget.classList.add('disable');
+    document.querySelector('.note').innerText = "Max";
+  }
+});
 var fileInput = document.getElementById('fileInput');
 fileInput.addEventListener('change', function (event) {
   fileArr.push(event.target.files[0]);
   filesToBlod(event.target.files[0]);
 });
 function upload() {
-  fileInput.click();
+  fileInput.addEventListener('click', handleEvent);
 }
 function filesToBlod(file) {
+  if (fileArr.length > 10) {
+    alert('Over limit');
+  }
   var reader = new FileReader();
   reader.readAsDataURL(file);
   reader.onload = function (e) {
-    fileBlodArr.push(e.target.result);
-    var fileDiv = document.createElement('div');
-    fileDiv.classList.add('drop-image-box');
+    newFunction();
+    function newFunction() {
+      if (fileArr.length <= 10) {
+        fileBlodArr.push(e.target.result);
+        var fileDiv = document.createElement('div');
+        fileDiv.id = 'drop-image-box';
+        fileDiv.classList.add('drop-image-box');
+        fileDiv.draggable = true; // 将元素设置为可拖拽
 
-    // 文件名称
-    var fileName = document.createElement('p');
-    fileName.classList.add('drop-image-box-title');
-    fileName.innerHTML = file.name;
-    fileName.title = file.name;
-    var deleteBtn = document.createElement('div');
-    deleteBtn.classList.add('drop-image-delete');
-    deleteBtn.innerText = 'Delete';
+        // 文件名称
+        var fileName = document.createElement('p');
+        fileName.classList.add('drop-image-box-title');
+        fileName.innerHTML = file.name;
+        fileName.title = file.name;
+        var deleteBtn = document.createElement('div');
+        deleteBtn.classList.add('drop-image-delete');
+        deleteBtn.innerText = 'Delete';
 
-    // 顯示照片略縮圖
-    var img = document.createElement('img');
-    img.classList.add('thumble');
-    img.src = e.target.result;
-    fileDiv.appendChild(img);
-    fileDiv.appendChild(deleteBtn);
-    fileDiv.appendChild(fileName);
-    dropImageList.appendChild(fileDiv);
-    deleteBtn.addEventListener('click', function () {
-      deleteFile(file);
-      dropImageList.removeChild(fileDiv);
-    });
+        // 顯示照片略縮圖
+        var img = document.createElement('img');
+        img.classList.add('thumble');
+        img.draggable = true;
+        img.src = e.target.result;
+        fileDiv.appendChild(img);
+        fileDiv.appendChild(deleteBtn);
+        fileDiv.appendChild(fileName);
+        dropImageList.appendChild(fileDiv);
+        imageQuantity.innerText = '(' + fileArr.length + '/10)';
+        deleteBtn.addEventListener('click', function () {
+          deleteFile(file);
+          dropImageList.removeChild(fileDiv);
+          imageQuantity.innerText = '(' + fileArr.length + '/10)';
+          droptarget.classList.remove('disable');
+          document.querySelector('.note').innerText = "Add Image";
+        });
+      }
+    }
   };
   reader.onerror = function () {
     switch (reader.error.code) {
       case '1':
-        alert('Can\'find the file');
+        alert("Can't find the file");
       case '2':
         alert('Security Error!');
       case '3':
         alert('Read interrupted');
       case '4':
-        alert('Can\'read the file');
+        alert("Can't read the file");
       case '5':
         alert('Read file failure');
     }
