+++ resources/js/backend/admin/upload.js
@@ -1,18 +1,42 @@
 let droptarget = document.getElementById('drop');
 let dropImageList = document.getElementById('dropImageList');
+let imageQuantity = document.querySelector('.count');
+let uploadButton = document.querySelector('#uploadButton');
 
 let fileArr = [];
 let fileBlodArr = [];
 
-function handleEvent(event) {
-  event.preventDefault();
-  if (event.type === 'drop') {
-    droptarget.classList.remove('active');
-    for (let file of event.dataTransfer.files) {
+uploadButton.addEventListener('change', (event) => {
+  let files = event.target.files;
+
+  // 遍历文件列表，将文件添加到 fileArr 数组中
+  for (let file of files) {
+    if (fileArr.length < 10) {
       fileArr.push(file);
       filesToBlod(file);
+      droptarget.classList.remove('disable');
+      document.querySelector('.note').innerText = "Add Image";
+    } else {
+      droptarget.classList.add('disable');
+      document.querySelector('.note').innerText = "Max";
     }
-  } else if (event.type === 'dropleave') {
+  }
+
+  // 重置 input 元素的值，清空选择的文件
+  event.target.value = '';
+})
+
+function handleEvent(event) {
+  event.preventDefault();
+  if (event.type === 'drop' || event.type === 'click') {
+    droptarget.classList.remove('active');
+    for (let file of event.dataTransfer.files) {
+      if(fileArr.length < 10){
+        fileArr.push(file);
+        filesToBlod(file); 
+      }
+    }
+  } else if (event.type === 'dragleave') {
     droptarget.classList.remove('active');
   } else {
     droptarget.classList.add('active');
@@ -24,6 +48,17 @@ droptarget.addEventListener('dragenter', handleEvent);
 droptarget.addEventListener('dragover', handleEvent);
 droptarget.addEventListener('drop', handleEvent);
 droptarget.addEventListener('dragleave', handleEvent);
+droptarget.addEventListener('click', ()=>{
+  if(fileArr.length < 10){
+    let uploadButton = document.getElementById('uploadButton');
+    uploadButton.click();
+    droptarget.classList.remove('disable');
+    document.querySelector('.note').innerText = "Add Image"
+  }else{
+    droptarget.classList.add('disable');
+    document.querySelector('.note').innerText = "Max"
+  }
+})
 
 let fileInput = document.getElementById('fileInput');
 fileInput.addEventListener('change', (event) => {
@@ -32,55 +67,74 @@ fileInput.addEventListener('change', (event) => {
 });
 
 function upload() {
-  fileInput.click();
+  fileInput.addEventListener('click', handleEvent);
 }
 
 function filesToBlod(file) {
+
+  if (fileArr.length > 10) {
+    alert('Over limit');
+  }
+
   let reader = new FileReader();
   reader.readAsDataURL(file);
 
   reader.onload = (e) => {
-    fileBlodArr.push(e.target.result);
+    newFunction();
 
-    let fileDiv = document.createElement('div');
-    fileDiv.classList.add('drop-image-box');
+    function newFunction() {
+      if (fileArr.length <= 10) {
+        fileBlodArr.push(e.target.result);
 
-    // 文件名称
-    let fileName = document.createElement('p');
-    fileName.classList.add('drop-image-box-title');
-    fileName.innerHTML = file.name;
-    fileName.title = file.name;
+        let fileDiv = document.createElement('div');
+        fileDiv.id = 'drop-image-box';
+        fileDiv.classList.add('drop-image-box');
+        fileDiv.draggable = true; // 将元素设置为可拖拽
 
-    let deleteBtn = document.createElement('div');
-    deleteBtn.classList.add('drop-image-delete');
-    deleteBtn.innerText = 'Delete';
 
-    // 顯示照片略縮圖
-    let img = document.createElement('img');
-    img.classList.add('thumble');
-    img.src = e.target.result;
+        // 文件名称
+        let fileName = document.createElement('p');
+        fileName.classList.add('drop-image-box-title');
+        fileName.innerHTML = file.name;
+        fileName.title = file.name;
 
-    fileDiv.appendChild(img);
-    fileDiv.appendChild(deleteBtn);
-    fileDiv.appendChild(fileName);
-    dropImageList.appendChild(fileDiv);
+        let deleteBtn = document.createElement('div');
+        deleteBtn.classList.add('drop-image-delete');
+        deleteBtn.innerText = 'Delete';
 
-    deleteBtn.addEventListener('click', () => {
-      deleteFile(file);
-      dropImageList.removeChild(fileDiv);
-    });
+        // 顯示照片略縮圖
+        let img = document.createElement('img');
+        img.classList.add('thumble');
+        img.draggable = true;
+        img.src = e.target.result;
+
+        fileDiv.appendChild(img);
+        fileDiv.appendChild(deleteBtn);
+        fileDiv.appendChild(fileName);
+        dropImageList.appendChild(fileDiv);
+        imageQuantity.innerText = '(' + fileArr.length + '/10)';
+
+        deleteBtn.addEventListener('click', () => {
+          deleteFile(file);
+          dropImageList.removeChild(fileDiv);
+          imageQuantity.innerText = '(' + fileArr.length + '/10)';
+          droptarget.classList.remove('disable');
+          document.querySelector('.note').innerText = "Add Image";
+        });
+      }
+    }
   };
 
   reader.onerror = () => {
     switch (reader.error.code) {
       case '1':
-        alert('Can\'find the file');
+        alert("Can't find the file");
       case '2':
         alert('Security Error!');
       case '3':
         alert('Read interrupted');
       case '4':
-        alert('Can\'read the file');
+        alert("Can't read the file");
       case '5':
         alert('Read file failure');
     }
